requirements.txt
click==7.1.2
mysql-connector-python==8.0.29
protobuf==4.21.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.20.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
